requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
PyJWT==2.10.1
reportlab==4.4.9
sqlparse==0.5.5
tzdata==2025.3
